+++ src/compiler/emitter/lines_1000_end.rs
@@ -5,6 +5,7 @@ use std::convert::TryInto;
 use std::iter::FromIterator;
 use std::rc::Rc;
 
+use super::{brackets, PipelinePhase};
 use crate::{
     create_text_writer, factory, file_extension_is, get_emit_flags,
     get_emit_module_kind_from_module_and_target, get_literal_text, get_new_line_character,
@@ -18,336 +19,7 @@ use crate::{
     SyntaxKind, TextRange,
 };
 
-lazy_static! {
-    static ref brackets: HashMap<ListFormat, (&'static str, &'static str)> = create_brackets_map();
-}
-
-pub(crate) fn is_build_info_file(file: &str) -> bool {
-    file_extension_is(file, Extension::TsBuildInfo.to_str())
-}
-
-pub(crate) fn get_output_declaration_file_name<TGetCommonSourceDirectory: FnMut() -> String>(
-    input_file_name: &str,
-    config_file: &ParsedCommandLine,
-    ignore_case: bool,
-    get_common_source_directory: Option<&mut TGetCommonSourceDirectory>,
-) -> String {
-    unimplemented!()
-}
-
-pub(crate) fn get_common_source_directory_of_config(
-    command_line: &ParsedCommandLine,
-    ignore_case: bool,
-) -> String {
-    let options = &command_line.options;
-    let file_names = &command_line.file_names;
-    unimplemented!()
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-enum PipelinePhase {
-    Notification,
-    Substitution,
-    Comments,
-    SourceMaps,
-    Emit,
-}
-
-pub fn create_printer(
-    printer_options: PrinterOptions,
-    handlers: Option<Rc<dyn PrintHandlers>>,
-) -> Printer {
-    let handlers = handlers.unwrap_or_else(|| Rc::new(DummyPrintHandlers));
-    let printer = Printer::new(printer_options, handlers);
-    printer.reset();
-    printer
-}
-
-struct DummyPrintHandlers;
-
-impl PrintHandlers for DummyPrintHandlers {
-    fn is_on_emit_node_supported(&self) -> bool {
-        false
-    }
-}
-
 impl Printer {
-    pub fn new(printer_options: PrinterOptions, handlers: Rc<dyn PrintHandlers>) -> Self {
-        let extended_diagnostics = printer_options.extended_diagnostics == Some(true);
-        let new_line =
-            get_new_line_character(printer_options.new_line, Option::<fn() -> String>::None);
-        let module_kind = get_emit_module_kind_from_module_and_target(
-            printer_options.module,
-            printer_options.target,
-        );
-        let preserve_source_newlines = printer_options.preserve_source_newlines;
-        let bundle_file_info = if printer_options.write_bundle_file_info == Some(true) {
-            Some(BundleFileInfo {
-                sections: vec![],
-                sources: None,
-            })
-        } else {
-            None
-        };
-        let relative_to_build_info = if bundle_file_info.is_some() {
-            Some(Debug_.check_defined(printer_options.relative_to_build_info.clone(), None))
-        } else {
-            None
-        };
-        let record_internal_section = printer_options.record_internal_section;
-        Self {
-            printer_options,
-            handlers,
-            extended_diagnostics,
-            new_line,
-            module_kind,
-            current_source_file: RefCell::new(None),
-            bundled_helpers: RefCell::new(HashMap::new()),
-            node_id_to_generated_name: RefCell::new(HashMap::new()),
-            auto_generated_id_to_generated_name: RefCell::new(HashMap::new()),
-            generated_names: RefCell::new(HashSet::new()),
-            temp_flags_stack: RefCell::new(vec![]),
-            temp_flags: Cell::new(TempFlags::Auto),
-            reserved_names_stack: RefCell::new(vec![]),
-            reserved_names: RefCell::new(HashSet::new()),
-            preserve_source_newlines: Cell::new(preserve_source_newlines),
-            next_list_element_pos: Cell::new(None),
-            writer: RefCell::new(None),
-            own_writer: RefCell::new(None),
-            write: Cell::new(Printer::write_base),
-            is_own_file_emit: Cell::new(false),
-            bundle_file_info: RefCell::new(bundle_file_info),
-            relative_to_build_info,
-            record_internal_section,
-            source_file_text_pos: Cell::new(0),
-            source_file_text_kind: Cell::new(BundleFileSectionKind::Text),
-
-            source_maps_disabled: Cell::new(true),
-            source_map_generator: RefCell::new(None),
-            source_map_source: RefCell::new(None),
-            source_map_source_index: Cell::new(-1),
-            most_recently_added_source_map_source: RefCell::new(None),
-            most_recently_added_source_map_source_index: Cell::new(-1),
-
-            container_pos: Cell::new(-1),
-            container_end: Cell::new(-1),
-            declaration_list_container_end: Cell::new(-1),
-            current_line_map: RefCell::new(None),
-            detached_comments_info: RefCell::new(None),
-            has_written_comment: Cell::new(false),
-            comments_disabled: Cell::new(false),
-            last_substitution: RefCell::new(None),
-            current_parenthesizer_rule: RefCell::new(None),
-            // const { enter: enterComment, exit: exitComment } = performance.createTimerIf(extendedDiagnostics, "commentTime", "beforeComment", "afterComment");
-            parenthesizer: factory.with(|factory_| factory_.parenthesizer()),
-        }
-    }
-
-    fn maybe_current_source_file(&self) -> Option<Rc<Node>> {
-        self.current_source_file.borrow().clone()
-    }
-
-    fn maybe_writer(&self) -> Option<Rc<dyn EmitTextWriter>> {
-        self.writer.borrow().clone()
-    }
-
-    fn writer(&self) -> Rc<dyn EmitTextWriter> {
-        self.writer.borrow().clone().unwrap()
-    }
-
-    fn has_global_name(&self, name: &str) -> Option<bool> {
-        self.handlers.has_global_name(name)
-    }
-
-    fn on_emit_node(&self, hint: EmitHint, node: &Node, emit_callback: &dyn Fn(EmitHint, &Node)) {
-        if self.handlers.is_on_emit_node_supported() {
-            self.handlers.on_emit_node(hint, node, emit_callback)
-        } else {
-            no_emit_notification(hint, node, emit_callback)
-        }
-    }
-
-    fn is_on_emit_node_supported(&self) -> bool {
-        true
-    }
-
-    fn is_emit_notification_enabled(&self, node: &Node) -> Option<bool> {
-        self.handlers.is_emit_notification_enabled(node)
-    }
-
-    fn substitute_node(&self, hint: EmitHint, node: &Node) -> Option<Rc<Node>> {
-        Some(
-            self.handlers
-                .substitute_node(hint, node)
-                .unwrap_or_else(|| no_emit_substitution(hint, node)),
-        )
-    }
-
-    fn on_before_emit_node(&self, node: Option<&Node>) {
-        self.handlers.on_before_emit_node(node)
-    }
-
-    fn on_after_emit_node(&self, node: Option<&Node>) {
-        self.handlers.on_after_emit_node(node)
-    }
-
-    fn on_before_emit_node_array(&self, nodes: Option<&NodeArray>) {
-        self.handlers.on_before_emit_node_array(nodes)
-    }
-
-    fn on_after_emit_node_array(&self, nodes: Option<&NodeArray>) {
-        self.handlers.on_after_emit_node_array(nodes)
-    }
-
-    fn on_before_emit_token(&self, node: Option<&Node>) {
-        self.handlers.on_before_emit_token(node)
-    }
-
-    fn on_after_emit_token(&self, node: Option<&Node>) {
-        self.handlers.on_after_emit_token(node)
-    }
-
-    fn write(&self, text: &str) {
-        (self.write.get())(self, text);
-    }
-
-    fn bundle_file_info(&self) -> Ref<BundleFileInfo> {
-        Ref::map(self.bundle_file_info.borrow(), |bundle_file_info| {
-            bundle_file_info.as_ref().unwrap()
-        })
-    }
-
-    fn bundle_file_info_mut(&self) -> RefMut<BundleFileInfo> {
-        RefMut::map(self.bundle_file_info.borrow_mut(), |bundle_file_info| {
-            bundle_file_info.as_mut().unwrap()
-        })
-    }
-
-    fn enter_comment(&self) {
-        // unimplemented!()
-    }
-
-    fn exit_comment(&self) {
-        // unimplemented!()
-    }
-
-    fn emit_binary_expression(&self) {
-        unimplemented!()
-    }
-
-    pub fn print_node(
-        &self,
-        hint: EmitHint,
-        node: &Node,
-        source_file: &Node, /*SourceFile*/
-    ) -> String {
-        match hint {
-            EmitHint::SourceFile => {
-                Debug_.assert(is_source_file(node), Some("Expected a SourceFile node."));
-            }
-            EmitHint::IdentifierName => {
-                Debug_.assert(is_identifier(node), Some("Expected an Identifier node."));
-            }
-            EmitHint::Expression => {
-                Debug_.assert(is_expression(node), Some("Expected an Expression node."));
-            }
-            _ => (),
-        }
-        match node.kind() {
-            SyntaxKind::SourceFile => {
-                return self.print_file(node);
-            }
-            SyntaxKind::Bundle => {
-                return self.print_bundle(node);
-            }
-            SyntaxKind::UnparsedSource => {
-                return self.print_unparsed_source(node);
-            }
-            _ => (),
-        }
-        self.write_node(hint, node, Some(source_file), self.begin_print());
-        self.end_print()
-    }
-
-    pub fn print_list(
-        &self,
-        format: ListFormat,
-        nodes: &NodeArray,
-        source_file: &Node, /*SourceFile*/
-    ) -> String {
-        self.write_list(format, nodes, Some(source_file), self.begin_print());
-        self.end_print()
-    }
-
-    pub fn print_bundle(&self, bundle: &Node /*Bundle*/) -> String {
-        self.write_bundle(bundle, self.begin_print(), None);
-        self.end_print()
-    }
-
-    pub fn print_file(&self, source_file: &Node /*SourceFile*/) -> String {
-        self.write_file(source_file, self.begin_print(), None);
-        self.end_print()
-    }
-
-    pub fn print_unparsed_source(&self, unparsed: &Node /*UnparsedSource*/) -> String {
-        self.write_unparsed_source(unparsed, self.begin_print());
-        self.end_print()
-    }
-
-    pub fn write_node(
-        &self,
-        hint: EmitHint,
-        node: &Node,
-        source_file: Option<&Node /*SourceFile*/>,
-        output: Rc<dyn EmitTextWriter>,
-    ) {
-        let previous_writer = self.maybe_writer();
-        self.set_writer(Some(output), None);
-        self.print(hint, node, source_file);
-        self.reset();
-        *self.writer.borrow_mut() = previous_writer;
-    }
-
-    pub fn write_list(
-        &self,
-        format: ListFormat,
-        nodes: &NodeArray,
-        source_file: Option<&Node /*SourceFile*/>,
-        output: Rc<dyn EmitTextWriter>,
-    ) {
-        let previous_writer = self.maybe_writer();
-        self.set_writer(Some(output), None);
-        if source_file.is_some() {
-            self.set_source_file(source_file);
-        }
-        self.emit_list(Option::<&Node>::None, Some(nodes), format);
-        self.reset();
-        *self.writer.borrow_mut() = previous_writer;
-    }
-
-    fn get_text_pos_with_write_line(&self) -> usize {
-        self.writer()
-            .get_text_pos_with_write_line()
-            .unwrap_or_else(|| self.writer().get_text_pos())
-    }
-
-    fn update_or_push_bundle_file_text_like(
-        &self,
-        pos: isize,
-        end: isize,
-        kind: BundleFileSectionKind, /*BundleFileTextLikeKind*/
-    ) {
-        let mut bundle_file_info = self.bundle_file_info_mut();
-        let last = last_or_undefined(&bundle_file_info.sections);
-        if let Some(last) = last.filter(|last| last.kind() == kind) {
-            last.set_end(end);
-        } else {
-            bundle_file_info
-                .sections
-                .push(BundleFileSection::new_text_like(kind, None, pos, end));
-        }
-    }
-
     pub fn write_bundle(
         &self,
         bundle: &Node, /*Bundle*/
@@ -374,14 +46,14 @@ impl Printer {
         unimplemented!()
     }
 
-    fn begin_print(&self) -> Rc<dyn EmitTextWriter> {
+    pub(super) fn begin_print(&self) -> Rc<dyn EmitTextWriter> {
         self.own_writer
             .borrow_mut()
             .get_or_insert_with(|| Rc::new(create_text_writer(&self.new_line)))
             .clone()
     }
 
-    fn end_print(&self) -> String {
+    pub(super) fn end_print(&self) -> String {
         let own_writer = self.own_writer.borrow();
         let own_writer = own_writer.as_ref().unwrap();
         let text = own_writer.get_text();
@@ -389,7 +61,12 @@ impl Printer {
         text
     }
 
-    fn print(&self, hint: EmitHint, node: &Node, source_file: Option<&Node /*SourceFile*/>) {
+    pub(super) fn print(
+        &self,
+        hint: EmitHint,
+        node: &Node,
+        source_file: Option<&Node /*SourceFile*/>,
+    ) {
         if let Some(source_file) = source_file {
             self.set_source_file(Some(source_file));
         }
@@ -397,12 +74,12 @@ impl Printer {
         self.pipeline_emit(hint, node);
     }
 
-    fn set_source_file(&self, source_file: Option<&Node /*SourceFile*/>) {
+    pub(super) fn set_source_file(&self, source_file: Option<&Node /*SourceFile*/>) {
         *self.current_source_file.borrow_mut() =
             source_file.map(|source_file| source_file.node_wrapper());
     }
 
-    fn set_writer(
+    pub(super) fn set_writer(
         &self,
         writer: Option<Rc<dyn EmitTextWriter>>,
         source_map_generator: Option<Rc<dyn SourceMapGenerator>>,
@@ -410,13 +87,13 @@ impl Printer {
         *self.writer.borrow_mut() = writer;
     }
 
-    fn reset(&self) {
+    pub(super) fn reset(&self) {
         self.set_writer(
             None, None, //TODO this might be wrong
         );
     }
 
-    fn emit(&self, node: Option<&Node>) {
+    pub(super) fn emit(&self, node: Option<&Node>) {
         if node.is_none() {
             return;
         }
@@ -424,16 +101,16 @@ impl Printer {
         self.pipeline_emit(EmitHint::Unspecified, node);
     }
 
-    fn emit_expression(&self, node: &Node /*Expression*/) {
+    pub(super) fn emit_expression(&self, node: &Node /*Expression*/) {
         self.pipeline_emit(EmitHint::Expression, node);
     }
 
-    fn pipeline_emit(&self, emit_hint: EmitHint, node: &Node) {
+    pub(super) fn pipeline_emit(&self, emit_hint: EmitHint, node: &Node) {
         let pipeline_phase = self.get_pipeline_phase(PipelinePhase::Notification, emit_hint, node);
         pipeline_phase(self, emit_hint, node);
     }
 
-    fn get_pipeline_phase(
+    pub(super) fn get_pipeline_phase(
         &self,
         phase: PipelinePhase,
         emit_hint: EmitHint,
@@ -450,7 +127,7 @@ impl Printer {
         Debug_.assert_never(phase, None);
     }
 
-    fn pipeline_emit_with_hint(&self, hint: EmitHint, node: &Node) {
+    pub(super) fn pipeline_emit_with_hint(&self, hint: EmitHint, node: &Node) {
         if false {
             unimplemented!()
         } else {
@@ -458,7 +135,7 @@ impl Printer {
         }
     }
 
-    fn pipeline_emit_with_hint_worker(&self, mut hint: EmitHint, node: &Node) {
+    pub(super) fn pipeline_emit_with_hint_worker(&self, mut hint: EmitHint, node: &Node) {
         if hint == EmitHint::Unspecified {
             match node.kind() {
                 SyntaxKind::Identifier => return self.emit_identifier(node),
@@ -503,7 +180,11 @@ impl Printer {
         unimplemented!("hint: {:?}, kind: {:?}", hint, node.kind());
     }
 
-    fn emit_literal(&self, node: &Node /*LiteralLikeNode*/, jsx_attribute_escape: bool) {
+    pub(super) fn emit_literal(
+        &self,
+        node: &Node, /*LiteralLikeNode*/
+        jsx_attribute_escape: bool,
+    ) {
         let text = self.get_literal_text_of_node(node);
         if false {
             unimplemented!()
@@ -512,7 +193,7 @@ impl Printer {
         }
     }
 
-    fn emit_identifier(&self, node: &Node /*Identifier*/) {
+    pub(super) fn emit_identifier(&self, node: &Node /*Identifier*/) {
         let text_of_node = self.get_text_of_node(node, Some(false));
         if let Some(symbol) = node.maybe_symbol() {
             self.write_symbol(&text_of_node, &symbol);
@@ -521,7 +202,7 @@ impl Printer {
         }
     }
 
-    fn emit_property_signature(&self, node: &Node /*PropertySignature*/) {
+    pub(super) fn emit_property_signature(&self, node: &Node /*PropertySignature*/) {
         let node_as_property_signature = node.as_property_signature();
         self.emit_node_with_writer(
             Some(&*node_as_property_signature.name()),
@@ -531,12 +212,12 @@ impl Printer {
         self.write_trailing_semicolon();
     }
 
-    fn emit_call_signature(&self, node: &Node /*CallSignature*/) {
+    pub(super) fn emit_call_signature(&self, node: &Node /*CallSignature*/) {
         // unimplemented!()
         self.write_punctuation("TODO call signature");
     }
 
-    fn emit_type_reference(&self, node: &Node /*TypeReferenceNode*/) {
+    pub(super) fn emit_type_reference(&self, node: &Node /*TypeReferenceNode*/) {
         let node_as_type_reference_node = node.as_type_reference_node();
         self.emit(Some(&*node_as_type_reference_node.type_name));
         self.emit_type_arguments(
@@ -545,17 +226,17 @@ impl Printer {
         );
     }
 
-    fn emit_function_type(&self, node: &Node /*FunctionTypeNode*/) {
+    pub(super) fn emit_function_type(&self, node: &Node /*FunctionTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO function type");
     }
 
-    fn emit_constructor_type(&self, node: &Node /*ConstructorTypeNode*/) {
+    pub(super) fn emit_constructor_type(&self, node: &Node /*ConstructorTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO constructor type");
     }
 
-    fn emit_type_literal(&self, node: &Node /*TypeLiteralNode*/) {
+    pub(super) fn emit_type_literal(&self, node: &Node /*TypeLiteralNode*/) {
         self.write_punctuation("{");
         let flags = if true {
             ListFormat::SingleLineTypeLiteralMembers
@@ -570,12 +251,12 @@ impl Printer {
         self.write_punctuation("}");
     }
 
-    fn emit_array_type(&self, node: &Node /*ArrayTypeNode*/) {
+    pub(super) fn emit_array_type(&self, node: &Node /*ArrayTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO array type");
     }
 
-    fn emit_tuple_type(&self, node: &Node /*TupleTypeNode*/) {
+    pub(super) fn emit_tuple_type(&self, node: &Node /*TupleTypeNode*/) {
         self.emit_token_with_comment(
             SyntaxKind::OpenBracketToken,
             node.pos(),
@@ -603,7 +284,7 @@ impl Printer {
         );
     }
 
-    fn emit_union_type(&self, node: &Node /*UnionTypeNode*/) {
+    pub(super) fn emit_union_type(&self, node: &Node /*UnionTypeNode*/) {
         self.emit_list(
             Some(node),
             Some(&node.as_union_or_intersection_type_node().types()),
@@ -611,22 +292,22 @@ impl Printer {
         );
     }
 
-    fn emit_intersection_type(&self, node: &Node /*IntersectionTypeNode*/) {
+    pub(super) fn emit_intersection_type(&self, node: &Node /*IntersectionTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO intersection type");
     }
 
-    fn emit_conditional_type(&self, node: &Node /*ConditionalTypeNode*/) {
+    pub(super) fn emit_conditional_type(&self, node: &Node /*ConditionalTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO conditional type");
     }
 
-    fn emit_parenthesized_type(&self, node: &Node /*ParenthesizedTypeNode*/) {
+    pub(super) fn emit_parenthesized_type(&self, node: &Node /*ParenthesizedTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO parenthesized type");
     }
 
-    fn emit_type_operator(&self, node: &Node /*TypeOperatorNode*/) {
+    pub(super) fn emit_type_operator(&self, node: &Node /*TypeOperatorNode*/) {
         let node_as_type_operator_node = node.as_type_operator_node();
         self.write_token_text(
             node_as_type_operator_node.operator,
@@ -640,21 +321,21 @@ impl Printer {
         );
     }
 
-    fn emit_indexed_access_type(&self, node: &Node /*IndexedAccessType*/) {
+    pub(super) fn emit_indexed_access_type(&self, node: &Node /*IndexedAccessType*/) {
         // unimplemented!()
         self.write_punctuation("TODO indexed access type");
     }
 
-    fn emit_literal_type(&self, node: &Node /*LiteralTypeNode*/) {
+    pub(super) fn emit_literal_type(&self, node: &Node /*LiteralTypeNode*/) {
         self.emit_expression(&*node.as_literal_type_node().literal);
     }
 
-    fn emit_import_type(&self, node: &Node /*ImportTypeNode*/) {
+    pub(super) fn emit_import_type(&self, node: &Node /*ImportTypeNode*/) {
         // unimplemented!()
         self.write_punctuation("TODO import type");
     }
 
-    fn emit_token_with_comment<TWriter: FnMut(&str)>(
+    pub(super) fn emit_token_with_comment<TWriter: FnMut(&str)>(
         &self,
         token: SyntaxKind,
         mut pos: isize,
@@ -709,7 +390,7 @@ impl Printer {
         pos
     }
 
-    fn emit_node_with_writer(&self, node: Option<&Node>, writer: fn(&Printer, &str)) {
+    pub(super) fn emit_node_with_writer(&self, node: Option<&Node>, writer: fn(&Printer, &str)) {
         if node.is_none() {
             return;
         }
@@ -720,7 +401,10 @@ impl Printer {
         self.write.set(saved_write);
     }
 
-    fn emit_type_annotation<TNodeRef: Borrow<Node>>(&self, node: Option<TNodeRef /*TypeNode*/>) {
+    pub(super) fn emit_type_annotation<TNodeRef: Borrow<Node>>(
+        &self,
+        node: Option<TNodeRef /*TypeNode*/>,
+    ) {
         if let Some(node) = node {
             let node = node.borrow();
             self.write_punctuation(":");
@@ -729,7 +413,7 @@ impl Printer {
         }
     }
 
-    fn emit_type_arguments(
+    pub(super) fn emit_type_arguments(
         &self,
         parent_node: &Node,
         type_arguments: Option<&NodeArray /*<TypeNode>*/>,
@@ -742,7 +426,7 @@ impl Printer {
         );
     }
 
-    fn write_delimiter(&self, format: ListFormat) {
+    pub(super) fn write_delimiter(&self, format: ListFormat) {
         match format & ListFormat::DelimitersMask {
             ListFormat::None => (),
             ListFormat::CommaDelimited => {
@@ -756,7 +440,7 @@ impl Printer {
         }
     }
 
-    fn emit_list<TNode: Borrow<Node>>(
+    pub(super) fn emit_list<TNode: Borrow<Node>>(
         &self,
         parent_node: Option<TNode>,
         children: Option<&NodeArray>,
@@ -773,7 +457,7 @@ impl Printer {
         );
     }
 
-    fn emit_node_list<TNode: Borrow<Node>>(
+    pub(super) fn emit_node_list<TNode: Borrow<Node>>(
         &self,
         emit: fn(&Printer, Option<&Node>),
         parent_node: Option<TNode>,
@@ -856,51 +540,51 @@ impl Printer {
         }
     }
 
-    fn write_base(&self, s: &str) {
+    pub(super) fn write_base(&self, s: &str) {
         self.writer().write(s);
     }
 
-    fn write_string_literal(&self, s: &str) {
+    pub(super) fn write_string_literal(&self, s: &str) {
         self.writer().write_string_literal(s);
     }
 
-    fn write_symbol(&self, s: &str, sym: &Symbol) {
+    pub(super) fn write_symbol(&self, s: &str, sym: &Symbol) {
         self.writer().write_symbol(s, sym);
     }
 
-    fn write_punctuation(&self, s: &str) {
+    pub(super) fn write_punctuation(&self, s: &str) {
         self.writer().write_punctuation(s);
     }
 
-    fn write_trailing_semicolon(&self) {
+    pub(super) fn write_trailing_semicolon(&self) {
         self.writer().write_trailing_semicolon(";");
     }
 
-    fn write_keyword(&self, s: &str) {
+    pub(super) fn write_keyword(&self, s: &str) {
         self.writer().write_keyword(s);
     }
 
-    fn write_space(&self) {
+    pub(super) fn write_space(&self) {
         self.writer().write_space(" ");
     }
 
-    fn write_property(&self, s: &str) {
+    pub(super) fn write_property(&self, s: &str) {
         self.writer().write_property(s);
     }
 
-    fn increase_indent(&self) {
+    pub(super) fn increase_indent(&self) {
         self.writer().increase_indent();
     }
 
-    fn decrease_indent(&self) {
+    pub(super) fn decrease_indent(&self) {
         self.writer().decrease_indent();
     }
 
-    fn write_token_node(&self, node: &Node, writer: fn(&Printer, &str)) {
+    pub(super) fn write_token_node(&self, node: &Node, writer: fn(&Printer, &str)) {
         writer(self, token_to_string(node.kind()).unwrap());
     }
 
-    fn write_token_text<TWriter: FnMut(&str)>(
+    pub(super) fn write_token_text<TWriter: FnMut(&str)>(
         &self,
         token: SyntaxKind,
         mut writer: TWriter,
@@ -917,7 +601,7 @@ impl Printer {
         })
     }
 
-    fn get_text_of_node(&self, node: &Node, include_trivia: Option<bool>) -> String {
+    pub(super) fn get_text_of_node(&self, node: &Node, include_trivia: Option<bool>) -> String {
         if false {
             unimplemented!()
         } else if (is_identifier(node) || false) && true {
@@ -927,17 +611,17 @@ impl Printer {
         unimplemented!()
     }
 
-    fn get_literal_text_of_node(&self, node: &Node) -> Cow<'static, str> {
+    pub(super) fn get_literal_text_of_node(&self, node: &Node) -> Cow<'static, str> {
         let flags = GetLiteralTextFlags::None;
 
         get_literal_text(node, self.maybe_current_source_file(), flags)
     }
 
-    fn emit_leading_comments_of_position(&self, pos: isize) {
+    pub(super) fn emit_leading_comments_of_position(&self, pos: isize) {
         // unimplemented!()
     }
 
-    fn emit_trailing_comments_of_position(
+    pub(super) fn emit_trailing_comments_of_position(
         &self,
         pos: isize,
         prefix_space: Option<bool>,
@@ -947,7 +631,7 @@ impl Printer {
     }
 }
 
-fn create_brackets_map() -> HashMap<ListFormat, (&'static str, &'static str)> {
+pub(super) fn create_brackets_map() -> HashMap<ListFormat, (&'static str, &'static str)> {
     HashMap::from_iter(IntoIterator::into_iter([
         (ListFormat::Braces, ("{", "}")),
         (ListFormat::Parenthesis, ("(", ")")),
@@ -956,14 +640,14 @@ fn create_brackets_map() -> HashMap<ListFormat, (&'static str, &'static str)> {
     ]))
 }
 
-fn get_opening_bracket(format: ListFormat) -> &'static str {
+pub(super) fn get_opening_bracket(format: ListFormat) -> &'static str {
     brackets
         .get(&(format & ListFormat::BracketsMask))
         .unwrap()
         .0
 }
 
-fn get_closing_bracket(format: ListFormat) -> &'static str {
+pub(super) fn get_closing_bracket(format: ListFormat) -> &'static str {
     brackets
         .get(&(format & ListFormat::BracketsMask))
         .unwrap()
